# backend/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Kapital API Documentation",
    "version": "1.0.0",
    "description": "API documentation for Kapital backend services"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the product"
          },
          "code": {
            "type": "string",
            "description": "Product code"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "minAmount": {
            "type": "number",
            "description": "Minimum loan amount"
          },
          "maxAmount": {
            "type": "number",
            "description": "Maximum loan amount"
          },
          "repaymentTerms": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Available repayment terms in months"
          },
          "interestRate": {
            "type": "number",
            "description": "Interest rate percentage"
          },
          "eligibility": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Eligibility criteria"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the product was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the product was last updated"
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Product code"
          },
          "name": {
            "type": "string",
            "description": "Product name"
          },
          "description": {
            "type": "string",
            "description": "Product description"
          },
          "minAmount": {
            "type": "number",
            "description": "Minimum loan amount"
          },
          "maxAmount": {
            "type": "number",
            "description": "Maximum loan amount"
          },
          "repaymentTerms": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "Available repayment terms in months"
          },
          "interestRate": {
            "type": "number",
            "description": "Interest rate percentage"
          },
          "eligibility": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Eligibility criteria"
          },
          "lateFee": {
            "type": "number",
            "description": "Late payment fee percentage"
          },
          "originationFee": {
            "type": "number",
            "description": "Origination fee percentage"
          },
          "legalFee": {
            "type": "number",
            "description": "Legal fee amount"
          },
          "applicationFee": {
            "type": "number",
            "description": "Application fee amount"
          },
          "requiredDocuments": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Required document types"
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Product features"
          },
          "loanTypes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Available loan purpose types"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the product is active"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "Get all active products or a specific product by code",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "description": "Product code to filter by"
          }
        ],
        "responses": {
          "200": {
            "description": "List of active products or a single product if code is provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/products/{id}": {
      "patch": {
        "summary": "Update a product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a product",
        "tags": [
          "Products"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}